sync clocks

+++ src/common/utils.c
@@ -1,14 +1,15 @@
 #include <stdio.h>
+#include <stdint.h>
 #include <sys/time.h>
 
-long get_time()
+int64_t get_time()
 {
     struct timeval tv;
     gettimeofday(&tv, NULL);
-    return (tv.tv_sec * 1000) + (tv.tv_usec / 1000);
+    return ((int64_t)tv.tv_sec * 1000) + (tv.tv_usec / 1000);
 }
 
-int generate_unqiue_id()
+int32_t generate_unqiue_id()
 {
     static int id = 0;
     return id++;
